Add automation system abstraction
