+++ dynawo-extensions/dynawo-extensions-impl/src/main/java/com/powsybl/dynawo/extensions/impl/model/DynawoPhaseShifterIModelImpl.java
@@ -10,33 +10,13 @@ package com.powsybl.dynawo.extensions.impl.model;
 import com.powsybl.dynawo.extensions.api.model.DynawoPhaseShifterIModel;
 import com.powsybl.iidm.network.TwoWindingsTransformer;
 
-import java.util.ArrayList;
-import java.util.Collections;
-
 /**
  * @author Laurent Issertial {@literal <laurent.issertial at rte-france.com>}
  */
-public class DynawoPhaseShifterIModelImpl extends AbstractDynawoModel<TwoWindingsTransformer, DynawoPhaseShifterIModel> implements DynawoPhaseShifterIModel {
-
-    private final ArrayList<String> dynamicModelIdPerVariant;
+public class DynawoPhaseShifterIModelImpl extends AbstractDynawoAutomationSystemModelImpl<TwoWindingsTransformer, DynawoPhaseShifterIModel> implements DynawoPhaseShifterIModel {
 
     public DynawoPhaseShifterIModelImpl(TwoWindingsTransformer extendable, String modelName, String parameterSetId, String dynamicModelId) {
-        super(extendable, modelName, parameterSetId);
-        this.dynamicModelIdPerVariant = new ArrayList<>(Collections.nCopies(
-                getVariantManagerHolder().getVariantManager().getVariantArraySize(), null));
-        this.dynamicModelIdPerVariant.set(getVariantIndex(), dynamicModelId);
-        perVariantList.add(dynamicModelIdPerVariant);
-    }
-
-    @Override
-    public String getDynamicModelId() {
-        return dynamicModelIdPerVariant.get(getVariantIndex());
-    }
-
-    @Override
-    public DynawoPhaseShifterIModelImpl setDynamicModelId(String dynamicModelId) {
-        setAttribute(dynamicModelIdPerVariant, dynamicModelId, "dynamicModelId");
-        return this;
+        super(extendable, modelName, parameterSetId, dynamicModelId);
     }
 
     @Override
